Formatting

+++ game/src/main/kotlin/content/area/morytania/port_phasmatys/TorturedSoul.kt
@@ -6,11 +6,13 @@ import world.gregs.voidps.type.random
 class TorturedSoul : Script {
     init {
         npcDeath("tortured_soul") {
-            say(when(random.nextInt(3)) {
-                0 -> "Your soul will be forfeit for this, mortal!"
-                1 -> "I will return, mortal!"
-                else -> "You cannot kill the undead, mortal!"
-            })
+            say(
+                when (random.nextInt(3)) {
+                    0 -> "Your soul will be forfeit for this, mortal!"
+                    1 -> "I will return, mortal!"
+                    else -> "You cannot kill the undead, mortal!"
+                },
+            )
         }
     }
 }
